# src/types/Request/Admin.request.ts
export interface SaveReqAdmin {
  admin_name: String,
  admin_email:String, 
  admin_password: String,
  }


  export interface GetAdmin {
    admin_email:String, 
    admin_password:String, 
  }
  // export interface UpdateReqProduct {
  //   _id: string;
  //   p_name: String,
  //   p_price:String, 
  //   p_discription: String,
  // }
  // export interface Getproduct {
  //   id: string
  // }
  // export interface search {
  //   p_name: string
  // }
  // export interface DeleteProduct {
  //   id: string
  // }
  // export interface getProductPrice{
 
  //   p_price:String, 
    
  //   }
 