# src/types/Request/waiter.request.ts
export interface SaveReqWaiter {
  waiter_name: String,
  waiter_email:String, 
  waiter_password: String,
  }


export interface GetWaiter {
    waiter_email:String, 
    waiter_password:String, 
  }
  // export interface UpdateReqProduct {
  //   _id: string;
  //   p_name: String,
  //   p_price:String, 
  //   p_discription: String,
  // }
  // export interface Getproduct {
  //   id: string
  // }
  // export interface search {
  //   p_name: string
  // }
  // export interface DeleteProduct {
  //   id: string
  // }
  // export interface getProductPrice{
 
  //   p_price:String, 
    
  //   }
 